requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit
#faiss-cpu==1.7.4
#protobuf~=3.19.0
langchain_experimental
langchain_openai
sqlalchemy
sentence-transformers
snowflake